Fix PWA landscape mode and add custom icons
Both landscape and portrait modes were working in the browser, but when
you installed it as a PWA link from to android devices, it only worked
in portrait mode. This was just a setting in the manifest that I had to
fix to support any orientation rather than portrait only.

While I was at it, I created a basic apple logo and converted it into
icons for the browser and android launcher just to get a better idea of
how those work in PWAs.