Refactor google analytics
I'm just moving the google analytics javascript snippet out of the
template and into a component in order to 1. simplify the build (so we
don't have to remember to specify the custom template) and 2. make it
easier to wrap te entire component into a conditional call based on
users consent (it bothers me that most implementations download
analytics.js from google regardless of whether users have consented to
tracking or not).