Add analytic events for clicking and not clicking foods
I've refactored the routes a bit to include the step in the page url.
From an analytics perpective this probably doesn't actually make much
sense, but for now I'm trying out some different things just to see what
the consequences are. Our UI kind of turns the concept of content upside
down -- every choice includes two options and we want to see which ones
users are clicking on and which they aren't. For now I'm just including
custom events to track those events, but there might be a better way to
do it.

At the same time I'm just trying to get a better feel for how to use
(and how not to use) the preact framework. I have no prior react
experience, so I'm sure I'm following horrendous practices... It's a
learning process.