Refactored choice component
I struggle a little bit with the term choice. It's ambiguous because it
refers both to the set of things that one has to choose from AND to the
individual things that one might eventually choose -- as in Choice A,
Choice B, and the Choice of A or B. This only matters because it would
be nice to give our components meaningful names. I could use choice and
choice-list, but for now I've chosen to go with selection and option --
selection referring to a set of options. Once upon a time I considered
even calling our selections dilemmas, but that one is presented with two
undesirable options...

Mostly with this refactoring, I just wanted to separate out the
presentation of the food item itself from the logic related to handling
the particular act of selecting that food and differentiating it from
the alternatives. I also switched over from URL paths to query string
parameters for the various properties that we want to pass and added in
a branch property that's separate from the step property for navigating
our decision trees.