Adding new images and indexed json tree
With the angular version of feed me jefferson, I used json node trees to
define the decision tree. They were a bit of a pain to read though and
now that I've settled on a bit of a branch naming strategy, I'd like to
try out indexed trees were the tree is simply a map pointing terminal
branches to food ids. For non terminal branches we simply don't include
the node in the indexed tree leaving it up to the client to be smart
enough to know what to do when the node isn't there, and how to navigate
the branch ids.