Initial implementation of crystal bowl
This was just a quick and dirty half baked attempt to implement a
working crystal bowl UI.

The use of fetch for grabbing the decision tree json index has broken
the standard build that includes prerendering, but prerendering doesn't
necessarily make sense for us at this stage, so I've adjusted by simply
including --no-prerender in the build.

I want to clean up the implementation, but I'm still trying to get a
better feel for Props versus State in preact. This last little bit has
helped me get a better idea, but as a short cut, I took the horribly
lazy approach of passing the tree all the way down to the subcomponents
as a property. It's time to look into shared state management tools like
redux (or possibly unistore). This is something Angular handled
somewhat out of the box, but at quite a cost of framework overhead.