Implement expandable decision trees
This is mostly working now. I've broken the decision tree up so we can
somewhat lazily load it by eagerly loading partial branches that the
user is moving towards rather than loading the entire tree upfront.

The entire state is tract in a unistore now, so there are no more page
navigations when the user clicks on a food preference (meaning we'll
need to come up with a different way to manage stepping back). For now
there's no tracking hooked in, but that should be easy to add back in
with events and when the user hits a terminal node in the decision tree,
we just reroute to that page (meaning we clear out the full search
history state at that point, so that too will need to be fixed).