Implement shareable routes
For mobile versions this isn't so important, but for those using a
browser, it would be nice if the url shown in the address bar was
revisitable -- as in the user could copy and it from the address bar and
send it to a friend. Refreshing should also work, but really that's a
different issue because we also don't want to lose the users click
history if they jump over to a different page and come back, or hit the
refresh button.

The trick was getting all of this to work with partial decision tree
indexes. If somebody clicks on a link, they might start outside of the
scope of the bootstrap tree, or even closer to its edge on one of the
branches than we want them to get without expanding that branch. The new
logic is a bit of a mess, but it handles both of these situations
competently enough.