Update analytics event details
Changed the action, category and labels for dilemma click events. This
might make it easier to make sense of them in google analytics. I was
originally hesitant to make the specific food id a category (because
there are so many of them), but this seems to make more sense for
grouping than an arbitrary label like "selection". Prefer and decline
just seemed like nicer verbs for the action that users take on the
specific foods -- as in by clicking on X, I indicate that I prefer it
and by not clicking on Y, I have effectively declined it (I couldn't
find a better antonym for prefer -- reject seemed a bit harsh).

I also wrapped this in the start of an analytics consent indicator
because tests were failing with the calls to the undefined gtag
function. Ideally the tracker tag itself will get wrapped into that same
conditionality.

I also pulled in some linting config changes unintentionally -- I'm
about to start implementing random seeds for the dilemma food selection
and I plan to use bitwise operators, so I wanted to turn off that check
in the tslint config.