Implement randomized dilemma images
We don't want to always show the same images in or dilemmas or the users
will get bored (or maybe that wouldn't be such a bad thing?). Instead we
want to show a semi random image pulled from each of the branches shown
in the dillema. We'd like the random images to be as similar as possible
on all dimensions other than the dimension reflected in the dilemma, so
for each one we want to use the same random traversal of the tree.