Refactor transition between dilemma and recommendations
I've lazily been using hard routes/browser refreshes to transition
between the dilemma component and the final recommended food image. In
preparation for adding in some user feedback (thumbs up/thumbs down) on
the food recommendation, I just wanted to refactor it all so that we
don't have to lose our state during the transition. Eventually I'd like
to be able to present all of the foods that the user preferred as
alternatives if the user thumbs down the recommendation.