Build basic ui for procurement phase
This is just a little bit of a start for the feed-me/procurement phase.
We'll offer two simple buttons that just open up a google search link
for "pizza recipes" or "pizza delivery near me". The part that's missing
is a search term for each of the food items. It was getting to be time
to pull in our metadata anyway...