Implement basic attributions for images
This also included the start of some work for porting over general
metadata from feed me for the selected foods. For now I'm following the
same approach that we've used with partial tree indexes. For each index
we'll create a corresponding basket of metadata and we'll load it
proactively when we expand our branches.