Stitch in the procurement phase
Now when users click on the thumbs up to a recommendation, it will take
them to a page with links to order delivery, or search for recipes. It
works, but it's not without some glitchiness that still needs to be
fixed. Basically we need to start persisting the state when users leave
the site.