Refactor routes
I finally figured out that I should be using the route method instead of
window.push.state in order to trigger a routing event. I was having
trouble understanding why some things magically worked when the back
button was selected, but others not, and why we sometiems got into
broken states -- it was inconsistent routing... The router was getting
called on back button navigation, but not on window.push.state updates.

I think I need to rethink some of the routing so that we don't perform
redundant logic between the actions and the routes, but this should
already be half way there now.

It would be good to refactor the recommendations route so that it
includes a list of all of the recommendations rather than a single food
and then we just pop them off of the url with each thumb down rather
than relying completely on state. This should give better back button
navigation of the recommendations.