Refactor states and routes
I think I'm doing something very backwards with the new design, but
mainly I wanted to move us over to reproducible routes -- as in what you
see is what you ask for based on the url, so when it comes to dilemmas,
recommendations and procurment pages, the page should be fully
reproducible from the URL. This let's people share what they're seeing,
but also supports better navigation with back and forward buttons.

The way that we've accomplished this is by intercepting all of the
routes, and updating the state with the details passed in from the route
if it was available. The extra interception step means that we're
updating the state twice with ever request which doesn't make me feel
great. Also, the inconsistencies with which state variables are managed
by the routes and which aren't depending on which route your talking
about can kind of lead to some funny states. I think this is ok for now,
but I'm afraid it will lead to some frustration in the future.