Refactor tracker component
I'm still struggling to move the google analytics code out of the top
level html and into some library functions, but they haven't made it
easy. Instead I've just created some wrapper functions that type the
event arguments and include logic to suppress analytics if they're
turned off so that the logic is all centralized. The component class may
not be a great place to package those functions, but it's a work in
progress.