Disable analytics when the user has turned on "do not track"
A number of browsers support a privacy setting that will send request
headers indicating that the user wishes not to be tracked. Apparently
this is mostly ignored -- not only by users, but also by website
providers. We still need to implement a more in your face tracking
notification/consent mechanism (particularly for regions where this is
required for legal compliance), but in the meantime it seems like the
least we can do is disable analytics/tracking for users who have
explicitly asked us not to track them.