Persist state to local storage
When a user left the site and came back mid session, all of their
session history was forgotten even though they might have come right
back to the same page that they left from -- for instance if they had
clicked on three different food dilemmas and they were on branch 8, but
then clicked an attribution link to another page, or left to go to
another page and then came back via the back button, or even just hit
refresh, then we would lose all of the state that we had in memory --
including their choices that led up to this point, but also the state of
the expanded decision tree and loaded attribution data.

We can avoid losing all of this information by saving it to local
storage. We won't persist the details that are already available in the
route, but the choices, latest set of recommendations, the tree and
attribution data (basket) should all be persisted.

This implementation is a bit lazy/sloppy -- everything gets saved into
one variable. I'd like to make it a little nicer, but for now it's been
convenient just to use unissist. This is starting to show that
unistore/unissist might be a little too simplistic for our needs.