Implement minimalisting cookie consent popup
This provides a basic popup explaining that we use cookies and google
analytics for the purpose of improving the site and suggesting that
people say ok (with a nice bold ok button) but allowing them to opt out
with a subtle please don't button. The site is useable with the button
in the way, but only barely, so people should quickly accept or decline.
We won't start collecting any data until they've accepted though and
we'll remember their selection in the future. Also, for anybody with an
adblocker in place or "do not track" enabled in their browser, we won't
even bother to ask them, we'll just disable the cookies from the start.