Support custom food positioning
Most of our food images are relatively well centered, so defaulting to
centering the image, expanding it to fill up the space and overflow on
whatever dimension doesn't fit usually works. Some images are offset to
a direction though -- the strawberry smoothie (0000117) is a great
example. The image is landscape with mostly irrelevant background on the
right and the smoothie itself all the way over to the extreme left. On
some screens where the image is forced into a portrait shapped
container, the smoothie ends up completely off screen. As a basic fix to
this, I've added an optional position attribute to the food details that
can be set to top, bottom, left or right indicating which side of the
image should be favored. This maps directly to a css attribute that
determines how we display the image. For now I've only customized the
setting in the strawberry smoothie. In the future maybe we'll go further
toward setting actual percentages -- "position": "30% 70%" to identify
the ideal location for zooming into the food. I'll need to play around
more with that to figure out how best to use it though, and we'd
probably want a nice annotation tool for making it easier for moderators
to indicate where the food is in the image (or use some AI segmentation
tools to try to automate the process for us).