Fix bug with null/undefined detail
If the "detail" was null or undefined because it was either for a food
that hadn't yet been added to the basket, or it was one that wasn't in
the core basket and the rest of the locally saved basket hadn't yet been
"rehydrated", then it was causing the food component to break with a
null pointer error. As a result, even when the state was updated later
on and the attributions were available, the component had never rendered
in the first place, so it couldn't get updated.