Make sure baskets aren't loaded out of order
If a user deep links into a branch that has basket dependencies that
haven't been loaded -- like if they linked into branch 8 and baskets 1
and 4 hadn't yet been loaded, they should be loaded in the correct order
starting with the lowest number branch. Now we'll start loading any
ancestor branches that have baskets and apply them asyncronously in the
background until we're caught up with the current branch.