Start to move assets to configurable location
We want to use the same core logic for adams apple and feedme jefferson
(and potentially other skins), but be able to supply custom
themes/skins/models at build time via a build configuration. We'd also
like to be able to host the images and models (food baskets) at separate
domains (again, outside of the main static hosting domain) so that we
can update these separately from the static application build itself
(potentially to allow for dynamic real time updates).