Start customizing service worker by copying default over
The preact cli uses a custom service worker if it finds one in the
root source directory (not the root of the project). We want to
customize how we use workbox -- I'm not sure if this is quite the best
way to do it, but this is how I'm going to try the first time around.