Don't use opaque cache requests!
I was getting errors when I originally used cors fetch requests for the
just in time precaching of images, so I set the option to no-cors. It
turns out this was a bad idea as the opaque responses made it look like
we were storing a lot more data than we actually were, plus we couldn't
see if the responses were actually successful, so we might have ended up
caching bad responses. When I switched over to workbox's cache first
routing strategy, this also meant that we weren't even caching the
responses (but if you didn't actively clear out the cache they still
hung around from previous runs). Ultimately, we really want to run cache
first for images -- but only cache them upon successful responses (200,
not 0 which would included unknown opaque results). We have the power to
make cors requests since we control the CORS header in the server
responses -- and as far as I can tell, the errors had more to do with
previously cached workbox router responses. The annoying key to testing
this is to clear everything from memory, open the site, close your
browser, and then reopen the browser and site. Service workers really
are a major pain to test.

We also don't want to be caching the basket details -- partly since this
would be redundant to what we're already saving in internal storage, but
also because it could cause a lot of false positves/negatives in future
testing if we end up clearing out our internal storage and then
reloading from the cache (plus it could cause unexpected results during
offline mode).