Fix local storage persistence and timing
There were a few bugs in our implementation of unisisst for persistence
on local storage. First off, unisist was not automatically maintaining
the version for us -- so we were always going through the migration step
rather than the hydration step. Second, there was a null pointer issue
in our migration step meaning that we were never really going through
any step.

Now we've added the version on to the serialized set of data that gets
stored to local storage, so we should be able to compare it to the
version in the code properly.

Secondly, unisisst doesn't really offer a good hook for deferring any
activity until the state has been successfully loaded from local
storage. Lately, every time we hit the first page, we keep loading the
first branch because the default route redirects us to branch one which
loads basket 1 if it isn't already in set of loaded baskets. We've been
hitting this point before our persisted state has had a chance to load
from disk, so we keep loading it every time we revisit the site even if
we don't need to.

I've added a very hackish promise into the state implementation that
gets resolve in all of the hooks provided by unissist -- either the
hydration, migration and finally even the mapping hook. That last one
gets called when the data is saved to disk rather than when it's loaded
from disk, however if we happen to hit a bug in our own migration code
(like a null pointer), we could get kicked out of the hook before
resolving the promise. This isn't a great solution, but it's something
for now. It's also made it glaringly clear to me that I need to get
around to refactoring the dilemma routing flow sooner rather than later.