Eagerly prefetch images for upcoming dilemmas
Lazily precaching at the basket level hasn't worked out so well. Rather
than focusing on always having offline mode working, I think we need to
focus on having a better online experience at the moment. Precaching an
entire basket at a time has actually meant slower responsiveness for the
food images that the user should be seeing right now -- because those
requests are waiting on the rest of the basket to load. I'm ditching the
eager loading of all of the images for a basket for now and replacing it
with logic that simply prefetches the 4 images for the next two possible
dilemmas -- as in if the user picks option A, they would see foods A and
B for that option, but if they pick option B, they'd see foods A and B
for that option -- so we'll just eagerly prefetch AA, AB, BA and BB even
though we'll only need half of them (because we can't know until the
user picks which two we'll need). This comes with two advantages --
depending on how long it takes the user to click, the images will
already be loaded and since we're using the fetch method with cors, the
images should end up in our cache. As is, if we didn't prefetch them,
they wouldn't get cached since they're set up as background images and
background images don't have any way to load cross origin resources
other than as opaque requests.