Make core basket theme dependent
This took a while to figure out. Dynamic imports are also an interesting
option, but as much as we can have webpack bind them staticly into the
code, they still require using a promise which means changing the logic
to by async. In the long run, that might be a better option anyway, but
for now I didn't want to go through too big of a refactor and I was able
to get this working with a combination of webpack module aliasing and
updating the tsconfig -- it was important to do both, not just to fix
the editor issues (tsconfig always points to the default adams-apple
theme) but also when running the webpack builds. If the tsconfig path
isn't provided, the webpack module alias doesn't get used. I don't know
why... Similarly, if the webpack alias isn't included, webpack throws an
error because babel can't find the location specified in tsconfig. Seems
hacky, but it works for now.