Build karens kitchen theme with feed me jefferson assets
It's time to start reskinning adams apple with the feedme jefferson
theme, but before we go live with it we'll need to test that the whole
theming functionality actually works correctly. Eventually we'll
probably have a few different themes that are easily identifiable and
use them as our various testing environments -- so that nobody ever
mistakenly thinks they're looking at the wrong environment. The old
feedme-stage environment has already become something of a legacy
administrative content curation environment, so we can't reuse that as
our new staging environment, instead I've created a new firebase app
called karens kitchen and I'm using it to work on the feedme jefferson
theme brand assets. When we're happy with it we can rename it to the
feedme jefferson theme and replace it with some new karens kitchen
themed assets. This also provided a good opportunity to go through the
setup of a new environment from scratch and test out/update all of the
various notes.