Use feedme-trees dependency from github packages
I've been using npm link to include a the feedme-trees library
dependency without publishing it to any actual node package
repositories. Github recently started offering package management as an
alternative to hosting your package on the npmjs registry. As much as
I've been reluctant to publish to npmjs, I figured the source code is
already on github, so why not...