Add pagination for caption bubble content
This is still a mess, but I've kind of cleaned up the separation between
the talking head animation transitions and the caption bubble content.
Depending on the users screen size and font settings, we can't really be
sure that all of the text content will fit inside of a caption bubble
that doesn't completely cover the screen. It would be nice if we could
mostly fix the size of the caption bubble while dynamically paginating
the text content that goes in it.

The code in this commit attempts to accomodate that while also
refactoring the talking head animations a bit. While experimenting, I
went back and forth between css animations and transitions a bit. I've
settled on animations partly because they offer finer grained control,
but mostly because the animation end event identifies the name of the
animation that ended whereas the transition end event only identifies
the style that has completed transitioning, not the name of the
transition. On some of our animated tutorial screens, there could be a
number of animations and transitions taking place at the same time, so
it's important that any updates that are triggered by an animation end
only respond to one animation that should be triggering them, and not
all the other ones that will also trigger the same type of event.