Implement user options in chef dialogues
This is very much a bunch of spaghetti code at the moment, but we'd
really like to get something working out the door so I'll leave it at
this for the most part for the mvp.

I've put together the basic skeleton for the welcome and tutorial
dialogue, but the actual script could use a bit of work. I also need to
update the background animation for the intro routes and clean up the
actual styling for the chef and captions, but I want to get feedback on
those before refining them too much more.