Add a menu, fix tutorial link routes bug
If you went directly to the intro route, when it started the chef, the
routes running in the background weren't working for some reason. This
wasn't a problem for first time welcome dialogues though. I don't really
know why, but kicking off the dialogue directly seemed to fix the
problem that calling it from the route was causing.