Clean up links, fix overflow
I've switched the nav links to use Link from the router since it
automatically takes care of shutting off the default browser navigation
and using the router instead. As for links to external pages, (such as
attributions) I went back and made sure they all open up content in new
tabs.

The second step of the tutorial was causing the page to be larger than
the screen on portrait devices (phones) which created a strange effect,
so I've fixed that.